refactor: move hooks to app component dir
